pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = 'nomad-normalizer-plugin-atomic-descriptors'
version = '1.0'
description = 'SOAP nomad plugin for NOMAD.'
readme = "README.md"
authors = [{ name = "The NOMAD Authors" }]
license = { text = "Apache-2.0" }
dependencies = [
    'quippy-ase==0.9.14',
    "nomad-lab[infrastructure]@git+https://github.com/nomad-coe/nomad.git@develop",
    "nomad-schema-plugin-run@git+https://github.com/nomad-coe/nomad-schema-plugin-run.git@develop",
]

[project.urls]
homepage = "https://github.com/nomad-coe/nomad-normalizer-plugin-atomic-descriptors"

[project.optional-dependencies]
dev = [
    'mypy==1.0.1',
    'pytest==3.10.0',
    'pytest-timeout==1.4.2',
    'pytest-cov==2.7.1',
    'ruff==0.1.4'
]

tests = [
]

[tool.ruff]
include = ["atomicdescriptorsnormalizer/*.py", "tests/*.py"]
select = [
    "E", # pycodestyle
    "W", # pycodestyle
    "PL", # pylint
]
ignore = [
    "E501", # Line too long ({width} > {limit} characters)
    "E701", # Multiple statements on one line (colon)
    "E731", # Do not assign a lambda expression, use a def
    "E402",  # Module level import not at top of file
    "PLR0911", # Too many return statements
    "PLR0912", # Too many branches
    "PLR0913", # Too many arguments in function definition
    "PLR0915", # Too many statements
    "PLR2004", # Magic value used instead of constant
    "PLW0603", # Using the global statement
    "PLW2901", # redefined-loop-name
    "PLR1714", # consider-using-in
    "PLR5501", # else-if-used
]
fixable = ["ALL"]

[tool.setuptools.packages.find]
include = ["atomicdescriptorsnormalizer*"]

[tool.setuptools.package-data]
atomicdescriptorsnormalizer = ['*/nomad_plugin.yaml']
